build: rename PreferencesManager to Preferences
